+++ droidlet/dialog/craftassist/dialogue_objects/__init__.py
@@ -1,13 +1,3 @@
 from .mc_dialogue_object import MCBotCapabilities
-from .mc_interpreter import MCInterpreter
-from .dummy_interpreter import DummyInterpreter
-from .get_memory_handler import MCGetMemoryHandler
-from .put_memory_handler import PutMemoryHandler
 
-__all__ = [
-    MCBotCapabilities,
-    MCGetMemoryHandler,
-    MCInterpreter,
-    DummyInterpreter,
-    PutMemoryHandler,
-]
+__all__ = [MCBotCapabilities]
