--- droidlet/dialog/robot/dialogue_objects/__init__.py
@@ -1,6 +1,0 @@
-from .loco_dialogue_object import LocoBotCapabilities
-from .loco_interpreter import LocoInterpreter
-from .get_memory_handler import LocoGetMemoryHandler
-from .put_memory_handler import PutMemoryHandler
-
-__all__ = [LocoBotCapabilities, LocoGetMemoryHandler, PutMemoryHandler, LocoInterpreter]
