+++ droidlet/interpreter/condition_helper.py
@@ -3,7 +3,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 """
 from typing import Optional, Union
 from word2number.w2n import word_to_num
-from ...shared_data_structs import ErrorWithResponse
+from ..shared_data_structs import ErrorWithResponse
 from droidlet.interpreter.condition import (
     Condition,
     NeverCondition,
