+++ droidlet/interpreter/craftassist/condition_helper.py
@@ -1,8 +1,8 @@
 """
 Copyright (c) Facebook, Inc. and its affiliates.
 """
-from droidlet.dialog.dialogue_objects import ConditionInterpreter
-from droidlet.interpreter.craftassist.mc_stop_condition import AgentAdjacentStopCondition
+from droidlet.interpreter import ConditionInterpreter
+from .mc_stop_condition import AgentAdjacentStopCondition
 from .block_helpers import get_block_type
 
 # this will become unnecessary with distance between
