+++ droidlet/interpreter/craftassist/dummy_interpreter.py
@@ -3,12 +3,8 @@ Copyright (c) Facebook, Inc. and its affiliates.
 """
 from typing import Tuple, Dict, Any, Optional
 
-from droidlet.dialog.dialogue_objects import (
-    DialogueObject,
-    ReferenceObjectInterpreter,
-    FilterInterpreter,
-    interpret_reference_object,
-)
+from droidlet.dialog.dialogue_objects import DialogueObject
+from ..interpreter import ReferenceObjectInterpreter, FilterInterpreter, interpret_reference_object
 
 from .condition_helper import MCConditionInterpreter
 from .attribute_helper import MCAttributeInterpreter
