+++ droidlet/interpreter/craftassist/facing_helper.py
@@ -3,7 +3,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 """
 
 from droidlet.shared_data_structs import ErrorWithResponse
-from droidlet.dialog.dialogue_objects import interpret_relative_direction
+from droidlet.interpreter import interpret_relative_direction
 from word2number.w2n import word_to_num
