+++ droidlet/interpreter/filter_helper.py
@@ -15,10 +15,10 @@ from droidlet.memory.memory_filters import (
     RandomMemorySelector,
 )
 from droidlet.base_util import number_from_span
-from ...shared_data_structs import ErrorWithResponse
+from ..shared_data_structs import ErrorWithResponse
 from .location_helpers import interpret_relative_direction
 from .comparator_helper import interpret_comparator
-from .dialogue_object_utils import tags_from_dict
+from .interpreter_utils import tags_from_dict
 
 CARDINAL_RADIUS = 20
