+++ droidlet/interpreter/robot/__init__.py
@@ -1,0 +1,5 @@
+from .loco_interpreter import LocoInterpreter
+from .get_memory_handler import LocoGetMemoryHandler
+from .put_memory_handler import PutMemoryHandler
+
+__all__ = [LocoGetMemoryHandler, PutMemoryHandler, LocoInterpreter]
